third_party/blink/web_tests/media/media-fragments/TC0074-expected.txt


Title: TC0074
Fragment: 't=0:00:60'
Comment: UA knows that this is an invalid media fragment, so it will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
